Move vmap to its own file
