+++ src/lib.rs
@@ -8,3 +8,4 @@ pub mod point;
 pub mod polygon;
 pub mod triangle;
 pub mod vertex;
+pub mod vertex_map;
