+++ src/polygon.rs
@@ -1,4 +1,3 @@
-use std::collections::{hash_map, HashMap};
 use std::fs;
 use std::path::Path;
 
@@ -8,95 +7,12 @@ use crate::{
     point::Point,
     triangle::Triangle,
     vertex::{Vertex, VertexId},
+    vertex_map::VertexMap,
 };
 
 #[derive(Debug, PartialEq)]
 pub struct Polygon {
     vertex_map: VertexMap,
-}
-
-
-// TODO now that this is abstracted out like this, it makes me realize it
-// would make good sense to define a newtype wrapper around hashmap that
-// encapsulates this. It's also making me rethink vertex a bit. It's been
-// clumsy that I have to have the prev/next Option types, when really for
-// a polygon in the way I'm approaching it, those should _always_ be
-// defined for a valid polygon, you'll never have dangling vertices. So
-// perhaps it will make good sense then to not pass in a vec of vertices
-// to Polygon::new, but instead just points that store the x/y coordinates.
-// Then as you iterate over them, you can create Vertex instances that have
-// non-Option type prev/next, and an ID. The Point types could also have a
-// name which could be propagated to the vertex, so that it could serve as
-// an alias for look-up, which will be useful for tests/visualizations.
-// I think this solves most of my problems I've had, it gets rid of Option
-// so you don't have to unwrap a bunch of stuff, and the vmap operations
-// can enforce that prev/next are always defined and are valid IDs in the
-// map, so internally it could unwrap() on get operations while still
-// offering a public lookup that could optional.
-
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct VertexMap(HashMap<VertexId, Vertex>);
-
-
-impl VertexMap {
-    pub fn new(points: Vec<Point>) -> Self {
-        let mut vertex_map = HashMap::new();
-
-        let num_points = points.len();
-        let vertex_ids = (0..num_points)
-            .map(|_| VertexId::new(None))
-            .collect::<Vec<_>>();
-
-        for (i, point) in points.into_iter().enumerate() {
-            let prev_id = vertex_ids[(i + num_points - 1) % num_points];
-            let curr_id = vertex_ids[i];
-            let next_id = vertex_ids[(i + num_points + 1) % num_points];
-            let v = Vertex::new(point, curr_id, prev_id, next_id);
-            vertex_map.insert(curr_id, v);
-        }
-
-        Self(vertex_map)
-    }
-
-    pub fn get(&self, k: &VertexId) -> &Vertex {
-        // Unwrapping since this is for internal use only
-        // and it will be assumed that internally we only 
-        // operate on valid IDs in the map
-        self.0.get(k).unwrap()
-    }
-
-    pub fn get_mut(&mut self, k: &VertexId) -> &mut Vertex{
-        self.0.get_mut(k).unwrap()
-    }
-
-    pub fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    pub fn remove(&mut self, k: &VertexId) -> Vertex {
-        self.0.remove(k).unwrap()
-    }
-
-    pub fn values(&self) -> hash_map::Values<'_, VertexId, Vertex> {
-        self.0.values()
-    }
-
-    pub fn anchor(&self) -> &Vertex {
-        // TODO I'm not yet convinced this is something I want, ultimately
-        // need something to initiate algorithms in the vertex chain.
-        // Could consider only exposing keys and then having polygon gen
-        // an anchor
-        self.values().collect::<Vec<_>>()[0]
-    }
-
-    pub fn update_next(&mut self, k: &VertexId, next: &VertexId) {
-        self.get_mut(k).next = next.clone();
-    }
-
-    pub fn update_prev(&mut self, k: &VertexId, prev: &VertexId) {
-        self.get_mut(k).prev = prev.clone();
-    }
 }
